refactor: rentEntity 등 위치 다시 core로 변경
