+++ tracky-web/src/main/java/kernel360/trackyweb/rent/application/RentService.java
@@ -12,10 +12,10 @@ import org.springframework.transaction.annotation.Transactional;
 import kernel360.trackycore.core.common.api.ApiResponse;
 import kernel360.trackycore.core.common.api.PageResponse;
 import kernel360.trackycore.core.common.entity.CarEntity;
+import kernel360.trackycore.core.common.entity.RentEntity;
 import kernel360.trackycore.core.common.provider.CarProvider;
 import kernel360.trackycore.core.common.sse.GlobalSseEvent;
 import kernel360.trackycore.core.common.sse.SseEvent;
-import kernel360.trackyweb.common.entity.RentEntity;
 import kernel360.trackyweb.rent.application.dto.request.RentRequest;
 import kernel360.trackyweb.rent.application.dto.response.RentResponse;
 import kernel360.trackyweb.rent.domain.provider.RentDomainProvider;
