+++ tracky-web/src/main/java/kernel360/trackyweb/rent/application/dto/response/RentResponse.java
@@ -3,7 +3,7 @@ package kernel360.trackyweb.rent.application.dto.response;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import kernel360.trackyweb.common.entity.RentEntity;
+import kernel360.trackycore.core.common.entity.RentEntity;
 
 public record RentResponse(
 	String rent_uuid,
