+++ tracky-web/src/main/java/kernel360/trackyweb/rent/infrastructure/repository/RentDomainRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import kernel360.trackyweb.common.entity.RentEntity;
+import kernel360.trackycore.core.common.entity.RentEntity;
 
 @Repository
 public interface RentDomainRepository extends JpaRepository<RentEntity, Long> {
