+++ tracky-web/src/main/java/kernel360/trackyweb/rent/infrastructure/repository/RentRepositoryCustom.java
@@ -5,7 +5,7 @@ import java.time.LocalDateTime;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import kernel360.trackyweb.common.entity.RentEntity;
+import kernel360.trackycore.core.common.entity.RentEntity;
 
 public interface RentRepositoryCustom {
 	Page<RentEntity> searchByFilter(String rentUuid, String rentStatus, LocalDateTime rentDate, Pageable pageable);
